deprecate formatLocation in favor of SourceCodeLocation
