+++ archunit/src/main/java/com/tngtech/archunit/core/domain/SourceCodeLocation.java
@@ -18,11 +18,11 @@ package com.tngtech.archunit.core.domain;
 import java.util.Objects;
 
 import com.tngtech.archunit.PublicAPI;
+import com.tngtech.archunit.base.Optional;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.tngtech.archunit.PublicAPI.Usage.ACCESS;
-import static com.tngtech.archunit.core.domain.Formatters.formatLocation;
 
 /**
  * Location in the source code of an ArchUnit domain object.
@@ -40,15 +40,38 @@ import static com.tngtech.archunit.core.domain.Formatters.formatLocation;
  */
 @PublicAPI(usage = ACCESS)
 public final class SourceCodeLocation {
+    private static final String LOCATION_TEMPLATE = "(%s:%d)";
+
+    @PublicAPI(usage = ACCESS)
+    public static SourceCodeLocation of(JavaClass sourceClass) {
+        return new SourceCodeLocation(sourceClass, 0);
+    }
+
+    @PublicAPI(usage = ACCESS)
+    public static SourceCodeLocation of(JavaClass sourceClass, int lineNumber) {
+        return new SourceCodeLocation(sourceClass, lineNumber);
+    }
+
+    private static String formatLocation(JavaClass sourceClass, int lineNumber) {
+        Optional<String> recordedSourceFileName = sourceClass.getSource().isPresent()
+                ? sourceClass.getSource().get().getFileName()
+                : Optional.<String>absent();
+        String sourceFileName = recordedSourceFileName.isPresent() ? recordedSourceFileName.get() : guessSourceFileName(sourceClass);
+        return String.format(LOCATION_TEMPLATE, sourceFileName, lineNumber);
+    }
+
+    private static String guessSourceFileName(JavaClass location) {
+        while (location.getEnclosingClass().isPresent()) {
+            location = location.getEnclosingClass().get();
+        }
+        return location.getSimpleName() + ".java";
+    }
+
     private final JavaClass sourceClass;
     private final int lineNumber;
     private final String description;
 
-    SourceCodeLocation(JavaClass sourceClass) {
-        this(sourceClass, 0);
-    }
-
-    SourceCodeLocation(JavaClass sourceClass, int lineNumber) {
+    private SourceCodeLocation(JavaClass sourceClass, int lineNumber) {
         this.sourceClass = checkNotNull(sourceClass);
         this.lineNumber = lineNumber;
         checkArgument(lineNumber >= 0, "Line number must be non-negative but was " + lineNumber);
